fix: config panel
